@@ -6,7 +6,8 @@
   "types": "dist/index",
   "scripts": {
     "build": "tsc",
-    "test": "./node_modules/.bin/karma start --single-run --browsers Chrome"
+    "test": "ng test --watch=false --browsers Chrome",
+    "lint": "tslint \"src/**\""
   },
   "repository": {
     "type": "git",
@@ -24,40 +25,42 @@
   },
   "homepage": "https://github.com/Pragmatists/ng-test-runner#readme",
   "dependencies": {
-    "@types/sinon": "^2.3.6",
-    "lodash": "^4.17.4",
-    "protractor": "^5.2.0",
-    "sinon": "^4.0.1"
+    "lodash": "^4.17.11",
+    "protractor": "^5.4.1",
+    "sinon": "^5.1.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.8",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/compiler-cli": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
+    "@angular/cli": "7.1.2",
+    "@angular/common": "^7.1.2",
+    "@angular/compiler": "^7.1.2",
+    "@angular/compiler-cli": "^7.1.2",
+    "@angular/core": "^7.1.2",
+    "@angular/forms": "^7.1.2",
+    "@angular/http": "^7.1.2",
+    "@angular/platform-browser": "^7.1.2",
+    "@angular/platform-browser-dynamic": "^7.1.2",
+    "@angular/router": "^7.1.2",
+    "@angular-devkit/build-angular": "~0.10.0",
     "@types/es6-promise": "0.0.33",
-    "@types/jasmine": "2.5.38",
-    "@types/lodash": "^4.14.77",
-    "@types/node": "^8.0.45",
-    "jasmine-core": "^2.8.0",
-    "jasmine-spec-reporter": "^3.2.0",
-    "karma": "^1.7.1",
+    "@types/jasmine": "3.3.0",
+    "@types/lodash": "^4.14.118",
+    "@types/node": "^10.12.12",
+    "@types/sinon": "^2.3.7",
+    "rxjs": "^6.3.3",
+    "jasmine-core": "^3.2.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^3.1.3",
     "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.3.0",
-    "karma-jasmine": "^1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-typescript-preprocessor": "^0.3.1",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-junit-reporter": "^2.0.0",
+    "karma-typescript-preprocessor": "^0.4.0",
     "ssri": "^5.2.2",
-    "ts-node": "^3.3.0",
-    "tslint": "^5.8.0",
-    "typescript": "^2.6.2",
-    "zone.js": "^0.8.18"
+    "ts-node": "^4.1.0",
+    "tslint": "^5.11.0",
+    "typescript": "^3.2.2",
+    "zone.js": "^0.8.26"
   }
 }
