@@ -4,27 +4,17 @@
 module.exports = function (config) {
     config.set({
         basePath: '',
-        frameworks: ['jasmine', '@angular/cli'],
+        frameworks: ['jasmine', '@angular-devkit/build-angular'],
         plugins: [
             require('karma-jasmine'),
             require('karma-chrome-launcher'),
             require('karma-jasmine-html-reporter'),
             require('karma-coverage-istanbul-reporter'),
             require('karma-typescript-preprocessor'),
-            require('@angular/cli/plugins/karma')
+            require('@angular-devkit/build-angular/plugins/karma'),
         ],
         client:{
             clearContext: false // leave Jasmine Spec Runner output visible in browser
-        },
-        files: [
-
-            { pattern: './test/test.ts', watched: false }
-        ],
-        preprocessors: {
-            './test/test.ts': ['@angular/cli']
-        },
-        mime: {
-            'text/x-typescript': ['ts','tsx']
         },
         coverageIstanbulReporter: {
             reports: [ 'html', 'lcovonly' ],
